requirements.txt
cryptography>=41.0.0
gnupg>=2.3.2
click>=8.1.0
rich>=13.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tabulate>=0.9.0
colorama>=0.4.6
boto3>=1.26.0
paramiko>=3.0.0
jinja2>=3.1.0
